Menu Update vroom_vroom.py
I added a main menu and a "You Crashed!" ending screen. The settings and customization do not have any function yet. I also need to add a background to the main menu. I am going to add a score counter during the game and possibly a session-high score on the main menu.